+++ packages/server/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "__tests__/serverResource.ts"],
+  "include": ["src"],
   "compilerOptions": {
     // https://github.com/tsconfig/bases/blob/main/bases/node18.json
     "lib": ["es2023", "DOM", "DOM.Iterable", "DOM.AsyncIterable"],
