+++ packages/tests/server/createUntypedClient.test.ts
@@ -1,4 +1,4 @@
-import { ignoreErrors } from './___testHelpers';
+import { ignoreErrors } from '@trpc/server/__tests__/suppressLogs';
 import { createTRPCUntypedClient } from '@trpc/client';
 import type { Unsubscribable } from '@trpc/server/observable';
