+++ packages/tests/server/react/regression/issue-4486-initialData-types.test.tsx
@@ -1,5 +1,5 @@
-import { ignoreErrors } from '../../___testHelpers';
 import { getServerAndReactClient } from '../__reactHelpers';
+import { ignoreErrors } from '@trpc/server/__tests__/suppressLogs';
 import { initTRPC } from '@trpc/server';
 import { konn } from 'konn';
