+++ packages/tests/server/regression/issue-4360-useInfiniteQuery-placeHolderData.test.tsx
@@ -1,5 +1,5 @@
-import { ignoreErrors } from '../___testHelpers';
 import { getServerAndReactClient } from '../react/__reactHelpers';
+import { ignoreErrors } from '@trpc/server/__tests__/suppressLogs';
 import { initTRPC } from '@trpc/server';
 import { konn } from 'konn';
 import { z } from 'zod';
