+++ prettier.config.js
@@ -12,11 +12,11 @@ export default {
     // Ensure test helpers are sorted first so that polyfills are loaded
     '___',
     '__',
-    '^@trpc/server/__tests__(.*)$',
-    '^@trpc/client/__tests__(.*)$',
-    '^@trpc/react-query/__tests__(.*)$',
-    '^@trpc/tanstack-react-query/__tests__(.*)$',
-    '^@trpc/next/__tests__(.*)$',
+    '^@trpc/server/src/__tests__(.*)$',
+    '^@trpc/client/src/__tests__(.*)$',
+    '^@trpc/react-query/src/__tests__(.*)$',
+    '^@trpc/tanstack-react-query/src/__tests__(.*)$',
+    '^@trpc/next/src/__tests__(.*)$',
     '<THIRD_PARTY_MODULES>',
     '^[./]',
   ],
